requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
torch
scikit-learn
pycryptodome
psutil
streamlit
pillow
requests
numpy
pandas
matplotlib
seaborn
torch
torchvision
scikit-learn
pycryptodome
psutil
scikit-image
taming-transformers-rom1504
pytorch_lightning==1.9.4

